requirements/base.txt
bcrypt==3.1.4
blueprint==3.4.2
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-PyMongo==0.5.1
Flask-Script==2.0.6
Flask-WTF==0.14.2
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
Pillow==5.0.0
pycparser==2.18
pymongo==3.6.0
PyPDF2==1.26.0
python-libnmap==0.7.0
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
xhtml2pdf==0.2.1
